Split out assets code from agents file
