+++ src/input/asset.rs
@@ -1,5 +1,5 @@
 //! Code for reading [Asset]s from a CSV file.
-use crate::agent::Asset;
+use crate::asset::Asset;
 use crate::input::*;
 use crate::process::ProcessMap;
 use anyhow::{ensure, Context, Result};
